[WIP] Script dumping and charmap.txt
New tool (not built by default): phscrdump. Very incomplete, ugly,
buggy, and inefficient.

Tries to print a .phscr to a mostly readable format.

Currently reuses some of preproc's code directly, and only supports a
few opcodes.

Hopefully, it can be adapted to genenerate readable script files which
can be compiled with GAS+preproc like in the Pokémon decomps.

The charmap.txt is about 30% complete. There are ~1000 kanji in this
game which need to be manually identified from a massive column of 16x16
glyphs. And, of course, they aren't in order.

Perhaps someone who is proficient at Japanese could help with this?

preproc itself needs to be set to 16-bit chars, as the game uses fixed
16-bit codepoints.