Initial asset dumper commit
This version of the dumper is slow. It also touches the filesystem
more than it needs to and leaves temp files that can interfere with
the build if the timestamps line up perfectly.
I'm working on it, I swear...

This dumper is edited from extract-phoenix_data in
Missingmew/phoenixtools. Credits go to him and the original author.